requirements/base.txt
django==1.10.1
djangorestframework==3.4.7

